test: restructure by group
