+++ packages/core/test/options.test.js
