+++ packages/core/test/unicode.test.js
@@ -1,0 +1,31 @@
+"use strict";
+
+var path = require("path"),
+    
+    namer  = require("test-utils/namer.js"),
+    
+    Processor = require("../processor.js");
+
+describe("/processor.js", () => {
+    describe("unicode", () => {
+        var processor;
+        
+        beforeEach(() => {
+            processor = new Processor({
+                namer
+            });
+        });
+
+        it("should support unicode classes & ids", () =>
+            processor.file(
+                require.resolve("./specimens/unicode.css")
+            )
+            .then(() => processor.output({
+                to : "./packages/core/test/output/unicode.css"
+            }))
+            .then((output) =>
+                expect(output.css).toMatchSnapshot()
+            )
+        );
+    });
+});
